Share fused Qwen3.5 router top-k across LLM and VLM
